+++ mg_custom_nodes/KLL535_ComfyUI_Simple_Qwen3-VL-gguf_20260903/CHANGELOG.md
@@ -1,4 +1,5 @@
 # Nightly
+- Add speculative decoding
 - Add dynamic image, audio, video input
 - Add "user_prompt_template" input
 - Add "bypass" input